lolz_manager/requirements.txt
# Core UI (desktop manager)
PySide6>=6.6,<7.0

# HTTP / API
httpx>=0.27,<1.0
aiohttp>=3.9,<4.0

# Data & validation
pydantic>=2.5,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0,<2.0

# Database
SQLAlchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
# Для шифрованной БД (опционально, см. README):
# pysqlcipher3>=1.2

# Scheduling & queues
APScheduler>=3.10,<4.0

# Logging
loguru>=0.7,<1.0

# Reports export
openpyxl>=3.1,<4.0

# Crypto (для шифрования токена AES-256)
cryptography>=42.0

# Charts (вкладка "Статистика")
pyqtgraph>=0.13,<1.0

# Packaging (dev)
pyinstaller>=6.3

# ── Telegram Bot (Лемур Mini App) ──────────────────────────────────────────
aiogram>=3.10,<4.0
